corr

+++ src/parsing/token/tokenize_arg.c
@@ -6,7 +6,7 @@
 /*   By: rpepi <rpepi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/18 11:49:49 by pepi              #+#    #+#             */
-/*   Updated: 2024/07/25 14:08:29 by rpepi            ###   ########.fr       */
+/*   Updated: 2024/07/25 15:23:36 by rpepi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ void	tokenize_arg(t_env *env, char *content)
 {
 	int		type;
 	t_token	*token;
-	int		index;
+	size_t	index;
 
 	if (!content)
 	{
